backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
python-dotenv==1.0.0
pydantic==2.5.0
pydantic[email]==2.5.0
python-multipart==0.0.6
google-cloud-storage==2.10.0
google-cloud-vision==3.4.5
google-cloud-videointelligence==2.11.4
google-generativeai==0.3.2
Pillow==10.1.0